# api-docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "YukiBloom 网站的文章、图片、个人内容和站点页面 API。JSON 响应统一使用 code、message、data；成功 code 为 0，失败 code 与 HTTP 状态码一致。",
        "title": "YukiBloom 后端接口",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/admin/pages": {
            "get": {
                "description": "需要先通过管理员登录接口登录，结果包含草稿和已发布页面。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理页面"
                ],
                "summary": "查询后台页面列表",
                "responses": {
                    "200": {
                        "description": "页面列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresponse.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "尚未登录",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "500": {
                        "description": "查询失败",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    }
                }
            },
            "post": {
                "description": "需要登录和 CSRF 令牌。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理页面"
                ],
                "summary": "创建页面草稿",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录接口返回的 csrfToken",
                        "name": "X-CSRF-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "页面内容",
                        "name": "page",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sitecontent.Page"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "新建页面",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresponse.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/sitecontent.Page"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "页面内容无效",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "401": {
                        "description": "尚未登录",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "403": {
                        "description": "CSRF 校验失败",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "409": {
                        "description": "语言和链接标识已被占用",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "500": {
                        "description": "保存失败",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    }
                }
            }
        },
...
        "/api/v1/admin/pages/{id}/publish": {
            "post": {
                "description": "需要登录和 CSRF 令牌。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理页面"
                ],
                "summary": "发布页面",
                "parameters": [
                    {
                        "type": "string",
                        "description": "页面 UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "登录接口返回的 csrfToken",
                        "name": "X-CSRF-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "已发布页面",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresponse.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/sitecontent.Page"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "页面 ID 格式错误",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "401": {
                        "description": "尚未登录",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "403": {
                        "description": "CSRF 校验失败",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "404": {
                        "description": "页面不存在",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "500": {
                        "description": "发布失败",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/pages/{id}/unpublish": {
            "post": {
                "description": "需要登录和 CSRF 令牌。撤回后页面不再出现在公开接口。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理页面"
                ],
                "summary": "撤回页面",
                "parameters": [
                    {
                        "type": "string",
                        "description": "页面 UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "登录接口返回的 csrfToken",
                        "name": "X-CSRF-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "已撤回页面",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresponse.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/sitecontent.Page"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "页面 ID 格式错误",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "401": {
                        "description": "尚未登录",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "403": {
                        "description": "CSRF 校验失败",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "404": {
                        "description": "页面不存在",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "500": {
                        "description": "撤回失败",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    }
                }
            }
        },
...
        "/api/v1/site-content": {
            "get": {
                "description": "返回公开站点资料、导航、精选内容、公告、已审核友链及音乐信息。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "站点内容"
                ],
                "summary": "查询公开站点内容",
                "responses": {
                    "200": {
                        "description": "公开站点内容",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresponse.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/sitecontent.Content"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "查询失败",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apiresponse.Envelope": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "sitecontent.Page": {
            "type": "object",
            "properties": {
                "bodyMarkdown": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "locale": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
...
    }
}